Refactor Card component structure and improve imports across accessibility components; remove deprecated CardComponents and Card UI files.
