+++ components/index.tsx
@@ -1,7 +1,8 @@
 export { Badge } from './Badge';
+export { Card, CardProps } from './Card/Card';
 export { ThemedText } from './ThemedText';
 export { ThemedView } from './ThemedView';
 export { Button } from './ui/Button';
-export { Card, CardProps } from './ui/Card';
 export { Slider } from './ui/Slider';
 export { Switch } from './ui/Switch';
+
