+++ components/ui/index.tsx
@@ -1,7 +1,7 @@
 export { Button } from './Button';
-export { Card } from './Card';
 export { default as ColorPicker } from './ColorPicker';
 export { default as Dropdown } from './Dropdown';
 export { Picker } from './Picker';
 export { Slider } from './Slider';
 export { Switch } from './Switch';
+
